Change version to '0.1.6+nimbis.1' to indicate this is a Nimbis pkg.
This branch is identical to the upstream django-address master branch
except the version has the nimbis postfix.

The reason for adding the 'nimbis' postfix is so a separate package
can be uploaded to spatula and pulled into sites. Currently, even
though upstream is on version 0.1.6, there isn't a corresponding
package v0.1.6 released by the django-address maintainers.

This will allow us to use v0.1.6 in our sites, but still indicate
its a package uploaded by us and not django-address.