Re-apply removal of Media class from AddressWidget
We removed this back in 2018, but after upgrading django-address, we
lost the change because the AddressWidget class moved from forms.py to
widgets.py. The original commit message has been pasted below.

Original Commit ID: 32a6becebb6772cc34c400055a2d2e9de1e55373
Original Commit Message:

We are in the process decrecating django-address.  This was inspired by
it breaking instead of degrading gracefully.  This took out our sign up
form.

Reasoning for removal:

1) geocomplete (a Javascript dependency) is not longer maintained.

2) We can't easily convince it to no wreck the form it's associated
with when there is a failure.

3) We don't need addresses broken out in to constituent parts for us to
date (the other feature of this library).

We will need to keep the backend, or migrate it back to a simple
Charfield.  This commit simply removes the Javascript causes us the
pain.